sound: fix 1p multispawn ssm loading
this was affecting the game and watch fight at the end of all star mode. m-ex code was expecting the adjusted null fighter ID but a function was returning the vanilla null ID. this led to an out of bounds write to the SSM buffer sizes, causing pokemon.ssm buffer size to be 1 upon entering the fight, which led to heap size calculation errors down the line resulting in an assertion.